Fix stack order for modal image
The modal image div is currently put within the "middle" div. To make
its z-index take effect, move it outside so that it becomes a sibling of
the "middle" div and the footer. Increase its z-index to 10 (larger than
the value of floor buttons, which is 4).

Clean up the js code for modal image.