Fix display issues when scrolling
Several display problems will show up when the viewport height is less
than about 306px (for example when mobile device is in landscape mode):

1. The title "彼記-回憶大廳" goes out of the viewport.
2. The "2022 Peter Note" footer overlaps with "回彼記主頁".
3. The background image doesn't fill the whole page when scrolling to
   the bottom.

The core of all these problems lies in the fundamental design of the
page layout. In particular, the "middle" div is specified as
fixed-positioned, leaving its content to be likely to overflow on
screens with small height. Currently the "middle" div seems to be
scrollable only because of the "backdrop-filter" attribute set for the
body (see https://stackoverflow.com/questions/52937708).

In this commit we solve the problems altogether (hopefully). Major
changes are:

1. Applying "backdrop-filter" to the body is equivalent to applying
   "filter" to the background image. Therefore, move the background
   image to its own div to prevent affecting the positioning of the
   other elements in the page. Note that now the background image
   doesn't move when scrolling (no special reason for that, but I just
   think it's easier to implement).
2. IMO the footer doesn't have to be fixed-positioned, so move it
   within the "middle" div.
3. Allow scrolling the "container" divs (e.g. "about_container"), while
   making sure its orange background covers the whole viewport. For this
   to work, the "overflow-y" of the body needs to be set to "hidden"
   when the div is opened. After the change the "container" itself is
   fixed, but the content "container__content" is scrollable.