Improve error handling for timer
Since we are using an external service on opensheet.elk.sh, error
handling for the HTTP request is necessary. Handle the following
situations excplicitly in the code:

1. Data from the spreadsheet is invalid. This includes missing data and
   non-numeric data.
2. High response time from opensheet.elk.sh. Currently an asynchronous
   request is filed for each 5 seconds. If each request takes too long
   to process, we'll end up with multiple pending requests, which is
   undesirable. Solve the problem by adding a 1 second timeout for the
   request. According to my experiment, a succesful request takes less
   than 300ms.

In addition, it looks like we decided hide the timer before the game
starts. Therefore, remove the initial updateTimer() call, and use a
variable `endTimeLoaded` to avoid undefined reference error of `endHour`
and `endMinute`.

Finally, fix some style problems.