View images in a singler slider
Merge the 3 buttons "查看關卡位置", "查看玩家指引", "查看人物關係" into
one, "查看玩家指引", which opens up a modal image slider with the
original 3 images. The slider can be navigated by the left/right arrow
buttons, the left/right arrow keys on keyboard, or finger swiping on
mobile phones.

The slider is implemented by the Swiper API (swiperjs.com). The image
display is the same as before. For example, a vertical scrollbar will be
shown if the viewport height is too small. The maximum image width is
still 700px.